trial to add other objects
